correction boutton supprimer

+++ src/components/Home.vue
@@ -68,7 +68,7 @@
                                     <div><img class="rounded-circle mr-2" width="45" :src="post.profilePicture" alt=""></div>
                                     <div class="h5 m-0 "><a @click="goToProfile(post.UserId)">@{{post.username}}</a></div>
                                     <div class="text-muted ml-auto h7"><font-awesome-icon :icon="['far', 'clock']" />{{post.updatedAt | formatDate}}</div>
-                                    <div v-if="User == post.username || IsAdmin" class="ml-3"><button @click="deletePost(post.id)"> <font-awesome-icon :icon="['fas', 'times']" /></button></div>
+                                    <div v-if="User == post.username || IsAdmin" class="ml-3"><button class="del" @click="deletePost(post.id)"> <font-awesome-icon :icon="['fas', 'times']" /></button></div>
                     </div>
                     <div class="card-body">
                         <img v-if="post.attachment !== ''" class="aside img-responsive mb-4 w-100 h-auto mw-100 mh-100" :src="post.attachment"/>
@@ -179,7 +179,10 @@ export default {
     },
     async deletePost(id){
       try {
-        await this.DeletePost(id)
+        let reponse = window.confirm("Voulez vous vraiment supprimer ce message ?");
+        if(reponse){
+          await this.DeletePost(id)
+        }
       } catch (error){
         throw `${error}`
       }
@@ -205,6 +208,9 @@ label {
   cursor: pointer;
   border-radius:30px;
   margin: 10px;
+}
+.del{
+  border:none;
 }
 .unliked i{
